return an error if File.Close() ends in one
Errors can happen between the last call to File.Write and File.Close,
for example if any pending write operations fail.

Please note that errors can occur even after File.Close.
This is the case with omitted flushes and syncs. Usually this falls
into the responsibility of the kernel or system operator (adjusting
buffer sizes and so on), though.